added opentelemetry utils, separated utils
