+++ run.sh
@@ -1,2 +1,3 @@
 #! /bin/bash
 export PYTHONPATH=$PWD/src:$PYTHONPATH
+export OTEL_AGENT_CONFIG=/etc/otel/collector.yaml
