+++ src/models/__init__.py
@@ -1,0 +1,1 @@
+from src.models.model import *
