+++ src/utils/__init__.py
@@ -1,1 +1,1 @@
-from utils import utils
+from src.utils.common_utils import utils
